the program now codes the second character in the message
i am having trouble converting the ASCII code into the coded
characters. The code will print them if they are separated, but then I
would have to hard code the position of the characters when using the
text() function. I would rather be able to output a whole sentence in
one text() function. I have tried to mash the commands that I use to
convert and print the characters separately, but it does not work.